+++ src/game/tests.rs
@@ -282,6 +282,7 @@ mod vintage_cube_amped_raptor;
 mod vintage_cube_arrival;
 mod vintage_cube_artifacts;
 mod vintage_cube_barrowgoyf;
+mod vintage_cube_bitter_triumph;
 mod vintage_cube_bloodbraid_challenger;
 mod vintage_cube_bolass_citadel;
 mod vintage_cube_bountiful_landscape;